// src/generated/models/Tk_hotelbranchsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Tk_hotelbranchsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Tk_hotelbranchsstatecode = keyof typeof Tk_hotelbranchsstatecode;
export const Tk_hotelbranchsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Tk_hotelbranchsstatuscode = keyof typeof Tk_hotelbranchsstatuscode;

export interface Tk_hotelbranchsBase {
  importsequencenumber?: number;
  overriddencreatedon?: string;
  statecode: Tk_hotelbranchsstatecode;
  statuscode?: Tk_hotelbranchsstatuscode;
  timezoneruleversionnumber?: number;
  tk_branchaddress?: string;
  tk_branchcontactemail?: string;
  tk_branchcontactnumber?: string;
  "tk_BranchManager@odata.bind"?: string;
  tk_branchname: string;
  tk_branchref?: string;
  tk_hotelbranchid: string;
  utcconversiontimezonecode?: number;
}

export interface Tk_hotelbranchs extends Tk_hotelbranchsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  organizationidname: string;
  statecodename?: string;
  statuscodename?: string;
  tk_branchmanagername?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  organizationid?: object;
  _organizationid_value?: string;
  tk_branchmanager?: object;
  _tk_branchmanager_value?: string;
}